1998-03-10  Federico Mena Quintero  <federico@nuclecu.unam.mx>
	* gtools.c (message): Added return NULL.  This needs to be fixed.
	(translate_gdk_keysym_to_curses): FIXME: Replace KEY_BACKTAB by '\t'
	while KEY_BACKTAB is added somewhere else.
	(translate_gdk_keysym_to_curses): return -1 as last resort.

	* glayout.c (get_current_index): Add g_assert_not_reached() for
	fatal case.

	* gmenu.c (menubar_new): Cast g_menubar to unsigned long.

	* gwidget.c (x_create_input): Cast entry to GTK_OBJECT.

	* gkey.c (find_select_closure_callback): Fix = -> == typo.

1998-03-10  Federico Mena Quintero  <federico@nuclecu.unam.mx>

	* xslint.c (getch): Added missing "return".

	* utilunix.c (init_groups): Added parentheses around
	assignment/truth value.

	* boxes.c: #include <stdlib.h>

	* screen.c (string_inode): Cast fe->buf.st_ino to long to be
	consistent with sprintf format.

	* main.c (handle_args): Added parentheses around assignment/truth value.

+++ gnome/gmenu.c
@@ -65,7 +65,7 @@ WMenu *menubar_new (int y, int x, int cols, Menu menu [], int items)
     
     g_menubar = gtk_menu_bar_new ();
     gtk_widget_show (g_menubar);
-    menubar->widget.wdata = g_menubar;
+    menubar->widget.wdata = (unsigned long) g_menubar;
 
     for (i = 0; i < items; i++){
 	    GtkWidget *child;
